% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_NTE_wrapper.r
\name{assign_NTE_wrapper}
\alias{assign_NTE_wrapper}
\title{Assign NTE wrapper}
\usage{
assign_NTE_wrapper(genome, gtf_table, start_codons = c("ACG", "ATA",
  "ATC", "ATG", "ATT", "CTG", "GTG", "TTG"), stop_codons = (c("TAA",
  "TAG", "TGA")), max_nte_length = 300, genetic_code = GENETIC_CODE,
  context_span = c(-6, 4), truncations = TRUE,
  truncation_start_codons = "ATG", max_truncation_length = 50,
  cores = 1)
}
\arguments{
\item{genome}{BSgenome name, compatible with gtf}

\item{gtf_table}{gtf data-frame annotation}

\item{start_codons}{vector of triplets to be searched as start codons.}

\item{stop_codons}{vector of triplets to be searched as stop codons}

\item{max_nte_length}{maximum NTE length in aminoacids.}

\item{genetic_code}{GENETIC_CODE in biostrings genetic code format.}

\item{context_span}{two element vector, corresponding to start and end of the sequence relative to first position of start codon}

\item{truncations}{logical, if TRUE truncations will be searched for.}

\item{truncation_start_codons}{triplets to be searched for as truncations start codons, "ATG" by default.}

\item{max_truncation_length}{maximum truncation length in aminoacids.}

\item{cores}{number of cores to use. Passes argument to mcmapply() mc.cores argument.}
}
\value{
MitoCrypt_table of all isoforms called all genes in gtf table.
}
\description{
This function is multicore wrapper for gene_NTE_assigner().
}
\examples{

}
\seealso{
gene_NTE_assigner
}
\concept{MitoCrypt NTE_assign}
\keyword{NTE}
\keyword{extension}
\keyword{truncation}
