% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_generate.r
\name{block_generate}
\alias{block_generate}
\title{Block Generate}
\usage{
block_generate(gtf_table, transcript_id, fetch_feature = "CDS")
}
\arguments{
\item{gtf_table}{gtf annotation dataframe}

\item{transcript_id}{target gene}

\item{fetch_feature}{gene feature to be fetched. "CDS" by default.}
}
\value{
list of blockCount, blockSizes and blockStarts. See UCSC .bed file documentation for details.
}
\description{
This function takes gtf and returns .bed compatible blocks
}
\examples{

}
\keyword{bed}
\keyword{block}
