% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_tolower.r
\name{context_tolower}
\alias{context_tolower}
\title{Context tolower}
\usage{
context_tolower(context_sequence, context_span)
}
\arguments{
\item{context_sequence}{codon context character string object}

\item{context_span}{two element vector, corresponding to start and end of the sequence relative to first position of start codon}
}
\value{
start codon context character string
}
\description{
This function takes start character string codon context and returns character string where only start codon is displayed by capital letters.
}
\examples{
context_tolower('AAAAAAATGG', c(-6,5)) 

}
\keyword{codon}
\keyword{context}
\keyword{start}
