% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ntes.r
\name{extract_ntes}
\alias{extract_ntes}
\title{Extract NTEs}
\usage{
extract_ntes(min_mean, profile, ntes_input, max_nte)
}
\arguments{
\item{min_mean}{coverage threshold in RPKM}

\item{profile}{scikit-ribo A-site profile for processed gene}

\item{ntes}{ascending vector of NTE isoforms lengths.}
}
\value{
filtered NTEs vector
}
\description{
This function eliminates potential NTEs with ribosome profile coverage below given threshold
}
\examples{

}
\keyword{NTE}
\keyword{filter}
