% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequence.r
\name{fetch_sequence}
\alias{fetch_sequence}
\title{Fetch gene sequence based on gtf annotation and genome sequence}
\usage{
fetch_sequence(gtf_table, genome, transcript_id, output = c("DNAString",
  "indexes", "both"), fetch_feature = "CDS")
}
\arguments{
\item{gtf_table}{gtf annotation dataframe}

\item{genome}{BSgenome name, compatible with gtf}

\item{output}{output type. either "DNAstring" biostrings object, indexes, or both as a list.}

\item{fetch_feature}{gene feature to be fetched. "CDS" by default.}

\item{gene_id}{target gene}
}
\description{
This functions is wrapper of read.gtf function from refGenome.
}
\examples{
fetch_sequence()
}
\keyword{gtf}
\keyword{sequence}
