% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_NTE_assigner.r
\name{gene_NTE_assigner}
\alias{gene_NTE_assigner}
\title{Gene NTE assigner}
\usage{
gene_NTE_assigner(genome, gtf_table, start_codons = c("ACG", "ATA",
  "ATC", "ATG", "ATT", "CTG", "GTG", "TTG"), stop_codons = (c("TAA",
  "TAG", "TGA")), max_nte_length = 300, genetic_code = GENETIC_CODE,
  context_span = c(-6, 4), truncations = TRUE,
  truncation_start_codons = "ATG", max_truncation_length = 50, gene)
}
\arguments{
\item{genome}{BSgenome name, compatible with gtf}

\item{gtf_table}{gtf data-frame annotation}

\item{start_codons}{vector of triplets to be searched as start codons.}

\item{stop_codons}{vector of triplets to be searched as stop codons}

\item{max_nte_length}{maximum NTE length in aminoacids.}

\item{genetic_code}{GENETIC_CODE in biostrings genetic code format.}

\item{context_span}{two element vector, corresponding to start and end of the sequence relative to first position of start codon}

\item{truncations}{logical, if TRUE truncations will be searched for.}

\item{truncation_start_codons}{triplets to be searched for as truncations start codons, "ATG" by default.}

\item{max_truncation_length}{maximum truncation length in aminoacids.}

\item{gene}{transcript_id of gene of interest.}
}
\value{
MitoCrypt_table of all isoforms called for particular gene.
}
\description{
This function assigns potential extensions (and truncations) to gene of interest.
}
\examples{

}
\seealso{
assign_NTE_wrapper()
}
\concept{NTE_assign}
\keyword{NTE}
\keyword{extension}
\keyword{truncation}
