% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isoform_gtf.r
\name{get_isoform_gtf}
\alias{get_isoform_gtf}
\title{Get start/end modified isoform gtf}
\usage{
get_isoform_gtf(start, end, gtf_table, transcript_id,
  fetch_feature = "CDS")
}
\arguments{
\item{start}{new start of isoform}

\item{end}{new end of isoform}

\item{gtf_table}{gtf annotation dataframe}

\item{transcript_id}{target gene}

\item{fetch_sequence}{gene feature to be fetched. "CDS" by default.}
}
\value{
modified gtf entry
}
\description{
This function returns gtf entry modified with start or end location, taking into account gene structure (introns)
}
\examples{

}
