% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_expression.r
\name{isoform_expression}
\alias{isoform_expression}
\title{Isoform expression}
\usage{
isoform_expression(riboseq, query, mitocrypt, longest_nte,
  p_value = 0.01, ramp = 50, nte_cov = 1)
}
\arguments{
\item{riboseq}{riboseq_input.txt from scikit-ribo output}

\item{query}{gene_id}

\item{mitocrypt}{MitoCrypt table}

\item{longest_nte}{additional MitoCrypt table object, with one entry of longest searched for isoform for each gene.}

\item{p_value}{p value for permuatational t test comparing consecutive NTEs}

\item{ramp}{length of canonical ORF to be compared with its proximal isoform. In codons.}

\item{nte_cov}{minimum coverage of NTE isoform.}
}
\value{
MitoCrypt table of called isoforms for single gene
}
\description{
This function detects expressed NTE isoforms and returns absolute ribosome coverage
}
\examples{

}
\seealso{
isoform_expression_wrapper()
}
\keyword{detection}
\keyword{isoform}
\keyword{profiling}
\keyword{quantification}
