% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoform_expression_wrapper.r
\name{isoform_expression_wrapper}
\alias{isoform_expression_wrapper}
\title{Isoform expression wrapper}
\usage{
isoform_expression_wrapper(riboseq, mitocrypt, longest_nte,
  p_value = 0.01, ramp = 50, nte_cov = 1, cores = 1)
}
\arguments{
\item{riboseq}{riboseq_input.txt from scikit-ribo output}

\item{mitocrypt}{MitoCrypt table}

\item{longest_nte}{additional MitoCrypt table object, with one entry of longest searched for isoform for each gene.}

\item{p_value}{p value for permuatational t test comparing consecutive NTEs}

\item{ramp}{length of canonical ORF to be compared with its proximal isoform. In codons.}

\item{nte_cov}{minimum coverage of NTE isoform.}

\item{cores}{number of cores to use. Passes argument to mcmapply() mc.cores argument.}
}
\value{
MitoCrypt table of called isoforms for all genes present in longest_nte MitoCrypt_table
}
\description{
This function is multicore wrapper of isoform_expression()
}
\examples{

}
\seealso{
isoform_expression()
}
\keyword{detection}
\keyword{isoform}
\keyword{profiling}
\keyword{quantification}
