% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitofates_runner.r
\name{mitofates_runner}
\alias{mitofates_runner}
\title{Mitofates runner}
\usage{
mitofates_runner(sequence, mitofates_path = "MitoFates.pl", query_id,
  remove_intermediatory_files = TRUE, organism = c("fungi", "metazoa",
  "plant"))
}
\arguments{
\item{sequence}{aminoacid sequence character string}

\item{mitofates_path}{path to mitofates executable. By default "MitoFates.pl", works when "MitoFates.pl" is in environment $PATH. Perl dependencies must be in environment as well. Rstudio doesn't inherit global environment by default; "printenv | head -n -1 > .Renviron" should solve the issue.}

\item{query_id}{id of submitted query. Important when using multicore wrapper. Suggested query id is "paste(Gene.Name, NTE.Length, sep = '_')".}

\item{remove_intermediatory_files}{logical, if TRUE intermediatory files created in working directory are deleted. Strongly suggested to be set as TRUE, otherwise working directory might become cluttered with several thousands intermediatory files.}

\item{organism}{either "fungi" or "metazoa" or "plant"}
}
\value{
vector with Mitofates.Cleavage.Site and Mitofates.Probability
}
\description{
This function submits aminoacid sequence query to loacally installed MitoFates.
}
\examples{

}
\seealso{
mitofates_wrapper()

Other prediction_software: \code{\link{mitoprot_runner}},
  \code{\link{targetp_runner}}
}
\concept{prediction_software}
\keyword{mitofates}
\keyword{prediction}
