% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitofates_wrapper.r
\name{mitofates_wrapper}
\alias{mitofates_wrapper}
\title{Mitofates wrapper}
\usage{
mitofates_wrapper(MitoCrypt_table, mitofates_path = "MitoFates.pl",
  remove_intermediatory_files = TRUE, cores = 1,
  organism = c("fungi", "metazoa", "plant"))
}
\arguments{
\item{MitoCrypt_table}{MitoCrypt table dataframe}

\item{mitofates_path}{path to mitofates executable. By default "MitoFates.pl", works when "MitoFates.pl" is in environment $PATH. Perl dependencies must be in environment as well. Rstudio doesn't inherit global environment by default; "printenv | head -n -1 > .Renviron" should solve the issue.}

\item{remove_intermediatory_files}{logical, if TRUE intermediatory files created in working directory are deleted. Strongly suggested to be set as TRUE, otherwise working directory might become cluttered with several thousands intermediatory files.}

\item{cores}{number of cores to use. Passes argument to mcmapply() mc.cores argument.}

\item{organism}{either "fungi" or "metazoa" or "plant"}
}
\value{
vector with Mitofates.Cleavage.Site and Mitofates.Probability
}
\description{
This function is multicore wrapper for mitofates_runner().
}
\examples{

}
\seealso{
mitofates_runner()

Other prediction_software MitoCrypt: \code{\link{mitoprot_wrapper}},
  \code{\link{targetp_wrapper}}
}
\concept{prediction_software MitoCrypt}
\keyword{mitofates}
\keyword{parallel}
