% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoprot_runner.r
\name{mitoprot_runner}
\alias{mitoprot_runner}
\title{Mitoprot runner}
\usage{
mitoprot_runner(sequence, mitoprot_path = "mitoprot", query_id,
  Cleavage_site = TRUE, Mitoprot_probability = TRUE,
  remove_intermediatory_files = TRUE)
}
\arguments{
\item{sequence}{aminoacid sequence character string}

\item{mitoprot_path}{path to mitoprot executable. By default "mitoprot", works when "mitoprot" is in environment $PATH.}

\item{query_id}{id of submitted query. Important when using multicore wrapper. Suggested query id is "paste(Gene.Name, NTE.Length, sep = '_')".}

\item{Cleavage_site}{logical, if TRUE cleavage site is predicted, otherwise step is skipped.}

\item{Mitoprot_probability}{logical, if TRUE mitoprot probability is calculated, otherwise step is skipped.}

\item{remove_intermediatory_files}{logical, if TRUE intermediatory files created in working directory are deleted. Strongly suggested to be set as TRUE, otherwise working directory might become cluttered with several thousands intermediatory files.}
}
\value{
vector c(CleavageSite, MitoprotProbability)
}
\description{
This function submits aminoacid sequence query to loacally installed mitoprot.
}
\examples{

}
\seealso{
mitoprot_wrapper()

Other prediction_software: \code{\link{mitofates_runner}},
  \code{\link{targetp_runner}}
}
\concept{prediction_software}
\keyword{mitoprot}
\keyword{prediction}
