% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoprot_wrapper.r
\name{mitoprot_wrapper}
\alias{mitoprot_wrapper}
\title{Mitoprot wrapper}
\usage{
mitoprot_wrapper(MitoCrypt_table, mitoprot_path = "mitoprot",
  remove_intermediatory_files = TRUE, cores = 1)
}
\arguments{
\item{MitoCrypt_table}{MitoCrypt table dataframe}

\item{mitoprot_path}{path to mitoprot executable. By default "mitoprot", works when "mitoprot" is in environment $PATH.}

\item{remove_intermediatory_files}{logical, if TRUE intermediatory files created in working directory are deleted. Strongly suggested to be set as TRUE, otherwise working directory might become cluttered with several thousands intermediatory files.}

\item{cores}{number of cores to use. Passes argument to mcmapply() mc.cores argument.}
}
\value{
MitoCrypt_table with two added columns: MitoProt.Cleavage.Site and MitoProt.Probability
}
\description{
This function is multicore wrapper for mitoprot_runner().
}
\examples{

}
\seealso{
mitoprot_runner() mcmapply()

Other prediction_software MitoCrypt: \code{\link{mitofates_wrapper}},
  \code{\link{targetp_wrapper}}
}
\concept{prediction_software MitoCrypt}
\keyword{mitoprot}
\keyword{parallel}
