% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribosome_plot.r
\name{profile_supstarts}
\alias{profile_supstarts}
\title{profile supstarts}
\usage{
profile_supstarts(riboseq, mitocrypt, query_gene,
  alternative_name = NULL, print_pdf = F, custom_window = 300,
  mitocrypt_supported)
}
\arguments{
\item{riboseq}{riboseq_input.txt from scikit-ribo output}

\item{mitocrypt}{MitoCrypt table}

\item{query_gene}{id of gene of interest}

\item{alternative_name}{alternative name to be displayed as plot title}

\item{print_pdf}{logical, if TRUE pdf of plot will be printed in working directory}

\item{custom_window}{distance from first supported isoform to be plotted. In codons.}

\item{mitocrypt_supported}{MitoCrypt table of supported isoforms (output of isoform_expression_wrapper())}
}
\value{
plot of ribosome profile
}
\description{
This function plots ribosome profile and displays supported NTE isoforms as vertical lines
}
\examples{

}
\seealso{
isoform_expression_wrapper()
}
\keyword{profile}
\keyword{ribosome}
