% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf_as_data_frame.r
\name{read_gtf_as_data_frame}
\alias{read_gtf_as_data_frame}
\title{A read gtf function wrapper}
\usage{
read_gtf_as_data_frame(file)
}
\arguments{
\item{file}{gtf file path.}
}
\description{
This functions is wrapper of read.gtf function from refGenome.
}
\examples{
read_gtf_as_data_frame()
}
\keyword{gtf}
