% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp_runner.r
\name{targetp_runner}
\alias{targetp_runner}
\title{TargetP runner}
\usage{
targetp_runner(sequence, targetP_path = "targetp", query_id,
  remove_intermediatory_files = TRUE, organism = c("non-plant",
  "plant"))
}
\arguments{
\item{sequence}{aminoacid sequence character string}

\item{targetP_path}{path to targetP executable. By default "targetp", works when "targetp" is in environment $PATH.}

\item{query_id}{id of submitted query. Important when using multicore wrapper. Suggested query id is "paste(Gene.Name, NTE.Length, sep = '_')".}

\item{remove_intermediatory_files}{logical, if TRUE intermediatory files created in working directory are deleted. Strongly suggested to be set as TRUE, otherwise working directory might become cluttered with several thousands intermediatory files.}

\item{organism}{either "non-plant" (default) or "plant"}
}
\value{
vector with TargetP.Probability, TargetP.Localization and TargetP.Signal.Peptide.Probability
}
\description{
This function submits aminoacid sequence query to loacally installed TargetP.
}
\examples{

}
\seealso{
targetp_wrapper()

Other prediction_software: \code{\link{mitofates_runner}},
  \code{\link{mitoprot_runner}}
}
\concept{prediction_software}
\keyword{prediction}
\keyword{targetp}
