% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp_wrapper.R
\name{targetp_wrapper}
\alias{targetp_wrapper}
\title{TargetP wrapper}
\usage{
targetp_wrapper(MitoCrypt_table, targetP_path = "targetp",
  remove_intermediatory_files = TRUE, cores = 1,
  organism = c("non-plant", "plant"))
}
\arguments{
\item{MitoCrypt_table}{MitoCrypt table dataframe}

\item{targetP_path}{path to targetP executable. By default "targetp", works when "targetp" is in environment $PATH.}

\item{remove_intermediatory_files}{logical, if TRUE intermediatory files created in working directory are deleted. Strongly suggested to be set as TRUE, otherwise working directory might become cluttered with several thousands intermediatory files.}

\item{cores}{number of cores to use. Passes argument to mcmapply() mc.cores argument.}

\item{organism}{either "non-plant" (default) or "plant"}
}
\value{
vector with TargetP.Probability, TargetP.Localization and TargetP.Signal.Peptide.Probability
}
\description{
This function is multicore wrapper for targetp_runner().
}
\examples{

}
\seealso{
targetp_runner()

Other prediction_software MitoCrypt: \code{\link{mitofates_wrapper}},
  \code{\link{mitoprot_wrapper}}
}
\concept{prediction_software MitoCrypt}
\keyword{parallel}
\keyword{targetp}
